add targetsubject and handle same target feedback when target will be delete

+++ ehrest/src/main/java/com/everhomes/rest/user/FeedbackCommand.java
@@ -10,7 +10,7 @@ import com.everhomes.util.StringHelper;
  *  <li>contentCategory:内容类别：0-其它、1-产品bug、2-产品改进、3-版本问题;11-敏感信息、12-版权问题、13
  *         -暴力色情、14-诈骗和虚假信息、15-骚扰；21-谣言、22-恶意营销、23-诱导分享；31-政治</li>
  *  <li>contact:联系方式</li>
- *  <li>content:反馈标题</li>
+ *  <li>subject:反馈标题</li>
  *  <li>content:反馈内容</li>
  *  <li>proofResourceUri:图片链接</li>
  * </ul>
@@ -19,7 +19,7 @@ public class FeedbackCommand {
     private Byte feedbackType;
     private Byte targetType;
     private Long targetId;
-    private Integer contentCategory;
+    private Long contentCategory;
     private String contact;
     private String subject;
     private String content;
@@ -50,11 +50,11 @@ public class FeedbackCommand {
         this.targetId = targetId;
     }
 
-    public Integer getContentCategory() {
+    public Long getContentCategory() {
         return contentCategory;
     }
 
-    public void setContentCategory(Integer contentCategory) {
+    public void setContentCategory(Long contentCategory) {
         this.contentCategory = contentCategory;
     }
